Create Dockerfile
The Dockerfile is a script that defines the instructions to create a Docker image for the web server. It specifies the base image, copies necessary files, exposes the required port, and sets the command to run the Nginx server. This file automates the process of setting up the web server in a containerized environment.