Update index.html
The index.html file serves as the main static content for the web server. When a user navigates to the web server’s address, this HTML file is rendered by the browser. It's a simple, lightweight HTML document intended to demonstrate the functionality of the web server and Docker setup.